requirements.txt
pip install redis
pip install celery
django-celery-results
django_celery_beat